refactor: viewModel 별 Factory 분리
